pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-agent-tidb"
version = "0.1.0"
description = "RAG chatbot using TiDB Cloud VECTOR and DeepSeek via Ollama"
requires-python = ">=3.9"
authors = [{ name = "Your Name" }]
readme = "README.md"

[project.optional-dependencies]
# keep torch optional to allow CPU-only installs
full = [
  "sentence-transformers>=2.6.1",
  "mysql-connector-python>=9.0.0",
  "pandas>=2.2.2",
  "python-dotenv>=1.0.1",
  "ollama>=0.3.2",
]

[tool.setuptools]
packages = ["rag_agent"]
